# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kids Art Tales API
  description: API for children's drawing app with image-to-story, image-to-cartoon, and image-to-animation features
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/users/{id}:
    get:
      summary: Get user profile
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
    patch:
      summary: Update user profile
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                avatarUrl:
                  type: string
                language:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
  /api/subscriptions/{userId}/usage:
    get:
      summary: Get daily usage
      tags:
        - Subscriptions
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Daily usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  pictureBooks:
                    type: integer
                  animations:
                    type: integer
                  audioStories:
                    type: integer
                  lastResetDate:
                    type: string
                    format: date
                  isNewDay:
                    type: boolean
  /api/subscriptions/{userId}/check-quota:
    post:
      summary: Check if user can consume quota
      tags:
        - Subscriptions
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum: [pictureBooks, animations, audioStories]
                membership:
                  type: string
                  enum: [free, gold]
                  default: free
      responses:
        '200':
          description: Quota check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  canConsume:
                    type: boolean
                  current:
                    type: integer
                  limit:
                    type: string
                  remaining:
                    type: string
                  membership:
                    type: string
  /api/subscriptions/{userId}/consume:
    post:
      summary: Consume a quota item
      tags:
        - Subscriptions
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum: [pictureBooks, animations, audioStories]
                membership:
                  type: string
                  enum: [free, gold]
                  default: free
      responses:
        '200':
          description: Quota consumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  usage:
                    type: object
                  remaining:
                    type: object
        '403':
          description: Daily quota exceeded
  /api/subscriptions/{userId}/reset:
    post:
      summary: Reset usage (admin/testing)
      tags:
        - Subscriptions
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usage reset
components:
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
        age:
          type: integer
        membership:
          type: string
          enum: [free, gold]
        points:
          type: integer
        creations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: [audio, picture, animation, picture-book]
              title:
                type: string
              createdAt:
                type: string
                format: date-time
              mediaUrl:
                type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
        pendingLikeRewards:
          type: integer
...